addressing comments from @tomleb
